Move out the search things to the service
