+++ ui/packages/consul-ui/app/components/data-collection/index.js
@@ -1,78 +1,11 @@
 import Component from '@glimmer/component';
 import { inject as service } from '@ember/service';
 import { computed, get, action } from '@ember/object';
+import { alias } from '@ember/object/computed';
 import { tracked } from '@glimmer/tracking';
 import { sort } from '@ember/object/computed';
 import { defineProperty } from '@ember/object';
-import Fuse from 'fuse.js';
 
-class FuzzySearch {
-  constructor(items, options) {
-    this.fuse = new Fuse(items, {
-      shouldSort: false, // use our own sorting for the moment
-      threshold: 0.4,
-      keys: Object.keys(options.finders) || [],
-      getFn(item, key) {
-        return (options.finders[key](item) || []).toString();
-      },
-    });
-  }
-
-  search(s) {
-    return this.fuse.search(s).map(item => item.item);
-  }
-}
-
-class PredicateSearch {
-  constructor(items, options) {
-    this.items = items;
-    this.options = options;
-  }
-
-  search(s) {
-    const predicate = this.predicate(s);
-    // Test the value of each key for each object against the regex
-    // All that match are returned.
-    return this.items.filter(item => {
-      return Object.entries(this.options.finders).some(([key, finder]) => {
-        const val = finder(item);
-        if (Array.isArray(val)) {
-          return val.some(predicate);
-        } else {
-          return predicate(val);
-        }
-      });
-    });
-  }
-}
-class ExactSearch extends PredicateSearch {
-  predicate(s) {
-    s = s.toLowerCase();
-    return item =>
-      item
-        .toString()
-        .toLowerCase()
-        .indexOf(s) !== -1;
-  }
-}
-class RegExpSearch extends PredicateSearch {
-  predicate(s) {
-    let regex;
-    try {
-      regex = new RegExp(s, 'i');
-    } catch (e) {
-      // Return a predicate that excludes everything; most likely due to an
-      // eager search of an incomplete regex
-      return () => false;
-    }
-    return item => regex.test(item);
-  }
-}
-const searchables = {
-  fuzzy: FuzzySearch,
-  exact: ExactSearch,
-  regex: RegExpSearch,
-};
 export default class DataCollectionComponent extends Component {
   @service('filter') filter;
   @service('sort') sort;
@@ -80,19 +13,13 @@ export default class DataCollectionComponent extends Component {
 
   @tracked term = '';
 
+  constructor() {
+    super(...arguments);
+    this.searchableMap = this.searchService.searchables;
+  }
+
   get kind() {
     return this.args.type;
-  }
-
-  @computed('term', 'args.search')
-  get searchTerm() {
-    return this.term || this.args.search || '';
-  }
-
-  @action
-  search(term) {
-    this.term = term;
-    return this.items;
   }
 
   get searchMethod() {
@@ -103,15 +30,34 @@ export default class DataCollectionComponent extends Component {
     return this.args.filters.searchproperties;
   }
 
-  @computed('args{items,items.content}')
-  get content() {
-    // TODO: Temporary little hack to ensure we detect DataSource proxy
-    // objects but not any other special Ember Proxy object like ember-data
-    // things. Remove this once we no longer need the Proxies
-    if (this.args.items.dispatchEvent === 'function') {
-      return this.args.items.content;
+  @computed('term', 'args.search')
+  get searchTerm() {
+    return this.term || this.args.search || '';
+  }
+
+  @computed('kind', 'searchMethod', 'filtered', 'searchProperties')
+  get searchable() {
+    const Searchable =
+      typeof this.searchMethod === 'string'
+        ? this.searchableMap[this.searchMethod]
+        : this.args.searchable;
+    return new Searchable(this.filtered, {
+      finders: Object.fromEntries(
+        Object.entries(this.searchService.predicate(this.kind)).filter(([key, value]) => {
+          return typeof this.searchProperties === 'undefined'
+            ? true
+            : this.searchProperties.includes(key);
+        })
+      ),
+    });
+  }
+
+  @computed('kind', 'args.sort')
+  get comparator() {
+    if (typeof this.args.sort === 'undefined') {
+      return [];
     }
-    return this.args.items;
+    return this.sort.comparator(this.kind)(this.args.sort);
   }
 
   @computed('comparator', 'searched')
@@ -127,7 +73,7 @@ export default class DataCollectionComponent extends Component {
     return this.sorted;
   }
 
-  @computed('filtered', 'searchTerm', 'searchable')
+  @computed('searchTerm', 'searchable', 'filtered')
   get searched() {
     if (this.searchTerm === '') {
       return this.filtered;
@@ -135,38 +81,34 @@ export default class DataCollectionComponent extends Component {
     return this.searchable.search(this.searchTerm);
   }
 
-  @computed('kind', 'searchMethod', 'filtered', 'searchProperties')
-  get searchable() {
-    const searchable =
-      typeof this.searchMethod === 'string' ? searchables[this.searchMethod] : this.args.searchable;
-    return new searchable(this.filtered, {
-      finders: Object.fromEntries(
-        Object.entries(this.search.predicate(this.kind)).filter(([key, value]) => {
-          return typeof this.searchProperties === 'undefined'
-            ? true
-            : this.searchProperties.includes(key);
-        })
-      ),
-    });
-  }
-
   @computed('kind', 'content', 'args.filters')
   get filtered() {
+    // if we don't filter, return a copy of the content so we end up with what
+    // filter will return when filtering ED recordsets
     if (typeof this.args.filters === 'undefined') {
-      return this.content;
+      return this.content.slice();
     }
     const predicate = this.filter.predicate(this.kind);
     if (typeof predicate === 'undefined') {
-      return this.content;
+      return this.content.slice();
     }
     return this.content.filter(predicate(this.args.filters));
   }
 
-  @computed('kind', 'args.sort')
-  get comparator() {
-    if (typeof this.args.sort === 'undefined') {
-      return [];
+  @computed('args{items,items.content}')
+  get content() {
+    // TODO: Temporary little hack to ensure we detect DataSource proxy
+    // objects but not any other special Ember Proxy object like ember-data
+    // things. Remove this once we no longer need the Proxies
+    if (this.args.items.dispatchEvent === 'function') {
+      return this.args.items.content;
     }
-    return this.sort.comparator(this.kind)(this.args.sort);
+    return this.args.items;
+  }
+
+  @action
+  search(term) {
+    this.term = term;
+    return this.items;
   }
 }
