+++ ui/packages/consul-ui/app/search/predicates/health-check.js
@@ -2,32 +2,12 @@ const asArray = function(arr) {
   return Array.isArray(arr) ? arr : arr.toArray();
 };
 export default {
-  Name: (item, value) => {
-    return item.Name.toLowerCase().indexOf(value.toLowerCase()) !== -1;
-  },
-  Node: (item, value) => {
-    return item.Node.toLowerCase().indexOf(value.toLowerCase()) !== -1;
-  },
-  Service: (item, value) => {
-    const lower = value.toLowerCase();
-    return (
-      item.ServiceName.toLowerCase().indexOf(lower) !== -1 ||
-      item.ServiceID.toLowerCase().indexOf(lower) !== -1
-    );
-  },
-  CheckID: (item, value) => (item.CheckID || '').toLowerCase().indexOf(value.toLowerCase()) !== -1,
-  ID: (item, value) => (item.Service.ID || '').toLowerCase().indexOf(value.toLowerCase()) !== -1,
-  Notes: (item, value) =>
-    item.Notes.toString()
-      .toLowerCase()
-      .indexOf(value.toLowerCase()) !== -1,
-  Output: (item, value) =>
-    item.Output.toString()
-      .toLowerCase()
-      .indexOf(value.toLowerCase()) !== -1,
-  ServiceTags: (item, value) => {
-    return asArray(item.ServiceTags || []).some(
-      item => item.toLowerCase().indexOf(value.toLowerCase()) !== -1
-    );
-  },
+  Name: (item, value) => item.Name,
+  Node: (item, value) => item.Node,
+  Service: (item, value) => item.ServiceName,
+  CheckID: (item, value) => item.CheckID || '',
+  ID: (item, value) => item.Service.ID || '',
+  Notes: (item, value) => item.Notes,
+  Output: (item, value) => item.Output,
+  ServiceTags: (item, value) => asArray(item.ServiceTags || []),
 };
